fix(csa-process): classify fatal markers by content so sidecar broad markers stay stderr-scoped (fix #1652 FP)
Fixes #1652 follow-up for PR #1655 cumulative review HIGH (session 01KSRSJPXQXHG6ASP1HBVJB37C).

The previous tier split only protected the built-in default path. Production pre-exec always writes fatal_error_markers into the sidecar, and FatalErrorRegexes::custom routed every sidecar marker into all-channel matching, so broad HTTP/status defaults from config still matched output.log and tmux pane. This changes sidecar and built-in construction to use the same content classifier: bare HTTP/status markers become stderr-only and provider/quota markers remain all-channel.

Also keeps the existing liveness test fixture process from execing away its session path so the required csa-process package test is deterministic.
